+++ types/unpack.d.ts
@@ -1,0 +1,3 @@
+import type { Unpacked } from './index.cjs'
+
+export default function unpack(input?: string | null): Unpacked